Added Enchanter enchantment for Lecterns.
This enchantment allows the user to select whatever enchantment they want to receive by writing the enchantment name in a book and quill, placing it on the lectern, and taking it out again. This results in an enchanted book with the desired enchantment. The main drawback is that the level of the enchantment is randomized and does not benefit from bookshelf bonuses as enchantment tables do.