Fixed an issue where the Stream could return a null item stack while searching for non-enchanted armor in the Vibranium enchantment, throwing a NPE.
Previously the Stream would iterate over each item in the player's armor contents and check if the item does not have the Vibranium enchantment. This works fine while filtering because CustomEnchant#hasEnchantment checks for null values. However, a NPE would be thrown when the Stream returns a null value on Stream#findAny. Now the Stream only returns a boolean (removes the findAny NPE) that is only true if *all* armor is not null and has the Vibranium enchantment.